Rebase code to add dedicated UKF file
