+++ fault/src/multipleinertialUkf.py
@@ -4,7 +4,6 @@ import matplotlib.pyplot as plt
 import numpy as np
 import warnings
 warnings.filterwarnings("ignore", category=DeprecationWarning)
-from scipy.stats import chi2
 # The path to the location of Basilisk
 # Used to get the location of supporting data.
 from Basilisk import __path__
@@ -20,6 +19,8 @@ from Basilisk.fswAlgorithms import inertialUKF
 bskPath = __path__[0]
 fileName = os.path.basename(os.path.splitext(__file__)[0])
 
+from .ukf import compute_chisquare, configure_inertialattfilter
+
 from .plots import (
     plot_attitude_error,
     plot_filter_result_sigma,
@@ -30,73 +31,6 @@ from .plots import (
     plot_rw_speeds,
     plot_filter_chisquare
 )
-
-
-
-def compute_chisquare(dataFilterCov_S, dataFilterInno, threshold=1e12):
-    """
-    Computes the Mahalanobis distance (chi-square) for a sequence of innovation and covariance matrices,
-    and plots the resulting values if the covariance is well-conditioned.
-
-    Parameters:
-    - dataFilterCov_S: np.ndarray of shape (N, 9), flattened 3x3 covariance matrices
-    - dataFilterInno: np.ndarray of shape (N, 3), innovation vectors
-    - threshold: float, condition number threshold for filtering singular matrices
-
-    Returns:
-    - dataChiSquare: np.ndarray of Mahalanobis distances
-    """
-    dataChiSquare = []
-    for i in range(dataFilterInno.shape[0]):
-        cov_i = dataFilterCov_S[i, :].reshape(3, 3)
-        cond = np.linalg.cond(cov_i)
-        if cond < threshold:
-            mahalanobis = dataFilterInno[i, :].T @ np.linalg.inv(cov_i) @ dataFilterInno[i, :]
-            dataChiSquare.append(mahalanobis)
-    dataChiSquare = np.array(dataChiSquare)
-    return dataChiSquare
-
-def setup_inertialattfilter(filterObject):
-    filterObject.alpha = 0.02
-    filterObject.beta = 2.0
-    filterObject.kappa = 0.0
-    filterObject.switchMag = 1.2
-    filterObject.stateInit = [0.0, 0.0, 0.0, 0.0, 0.0, 0.0]
-    filterObject.covarInit = [1.0, 0.0, 0.0, 0.0, 0.0, 0.0,
-                              0.0, 1.0, 0.0, 0.0, 0.0, 0.0,
-                              0.0, 0.0, 1.0, 0.0, 0.0, 0.0,
-                              0.0, 0.0, 0.0, 0.1, 0.0, 0.0,
-                              0.0, 0.0, 0.0, 0.0, 0.1, 0.0,
-                              0.0, 0.0, 0.0, 0.0, 0.0, 0.1]
-    sigmaMrpSquare = (1E-3) ** 2
-    sigmaRateSquare = (5E-4) ** 2
-    qNoise = np.identity(6)
-    qNoise[0:3, 0:3] = qNoise[0:3, 0:3]*sigmaMrpSquare
-    qNoise[3:6, 3:6] = qNoise[3:6, 3:6]*sigmaRateSquare
-    filterObject.qNoise = qNoise.reshape(36).tolist()
-
-def configure_inertialattfilter(filterObject, config, measurement_message):
-    setup_inertialattfilter(filterObject)
-    vcMsg = config["vcMsg"]
-    rwStateEffector = config["rwStateEffector"]
-    inertialAttFilterRwParamMsg = config["inertialAttFilterRwParamMsg"]
-    gyroInMsg = config["gyroInMsg"]
-    st_cov = config["st_cov"]
-    # connect message to spacecraft and RW configurations
-    filterObject.massPropsInMsg.subscribeTo(vcMsg)
-    filterObject.rwSpeedsInMsg.subscribeTo(rwStateEffector.rwSpeedOutMsg)
-    filterObject.rwParamsInMsg.subscribeTo(inertialAttFilterRwParamMsg)
-    filterObject.gyrBuffInMsg.subscribeTo(gyroInMsg)
-    # setup measurement model in the filter
-    starTracker1 = inertialUKF.STMessage()
-    starTracker1.noise = [st_cov, 0.0, 0.0,
-                          0.0, st_cov, 0.0,
-                          0.0, 0.0, st_cov]
-    star_tracker_list = [starTracker1]
-    filterObject.STDatasStruct.STMessages = star_tracker_list
-    filterObject.STDatasStruct.numST = len(star_tracker_list)
-    # connect filter star tracker to the true inertial attitude measurement
-    filterObject.STDatasStruct.STMessages[0].stInMsg.subscribeTo(measurement_message)
 
 
 def multipleinertialUkf(show_plots=False):
